requirements.txt
# Core dependencies
llama-index-core>=0.10.0
llama-parse>=0.4.0
python-dotenv>=1.0.0
openai>=1.12.0
llama-index-llms-openai  # If using OpenAI chat models
llama-index-embeddings-openai  # For OpenAI embeddings
llama-index-vector-stores-qdrant  # For Qdrant vector store
qdrant-client>=1.7.0  # Qdrant client
numpy>=1.24.0  # Required for vector operations

# API dependencies
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.1  # For request/response models

# Optional but recommended
python-multipart>=0.0.6  # For handling form data
requests>=2.31.0  # For making HTTP requests